Extracted into NavigationAnimation component
